requirements.txt
docker
PyYAML
schedule